pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "NNIPdevelopement"
dynamic = ["version"]
description = "AiiDA workchain to generate data for training and testing of neural networks."
authors = [{name = "Nataliia Manko"}, {name = "Davide Bidoggia"}, {name = "Antimo Marrazzo"}, {name = "Maria Peressi"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 4 - Beta",
    "Framework :: AiiDA"
]
keywords = ["aiida", "workchain"]
requires-python = ">=3.7"
dependencies = [
    "aiida-core>=2.5,<3",
    "voluptuous",
    "ase",
    "numba"
]

[project.urls]
Source = "https://github.com/mankoNm/NNIPDevelopment"



[project.entry-points.'aiida.workflows']
"NNIPdevelopement.datageneration" = "NNIPdevelopement.datageneration.datageneration:DataGenerationWorkChain"

#[project.entry-points."aiida.cmdline.data"]
#"diff" = "aiida_diff.cli:data_cli"

#[tool.flit.module]
#name = "NNIPdevelopement"
